% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-library.R
\name{allow_outcome}
\alias{allow_outcome}
\title{Throw an error unless allowed outcome is specified (not exported - internal saproj only)}
\usage{
allow_outcome(comparison_outcome, allowed_outcome)
}
\arguments{
\item{comparison_outcome}{list: The output of \code{\link{compare_library_snapshot}}}

\item{allowed_outcome}{character: The allowed outcome name(s) from 
\code{\link{compare_library_snapshot}}}
}
\value{
Returns the result of compare_library_snapshot()
}
\description{
This is a helper function for use in snapshot_library() and restore_library().
It stops function execution for conditions where they shouldn't be run,
which is based on the value returned by compare_library_snapshot().
For example: one shouldn't (typically) run snapshot_library() unless "snapshot_behind".
}
\examples{
compare_library_snapshot() \%>\% allow_outcome("snapshot_behind")
compare_library_snapshot() \%>\% allow_outcome("library_behind")
}
\seealso{
Other internal helper functions: \code{\link{get_library_path}}
}
\concept{internal helper functions}
\keyword{internal}
