% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{archive_project}
\alias{archive_project}
\title{Archive the saproj settings in an "saproj" folder}
\usage{
archive_project(files = c(".Rprofile", "snapshot-library.csv"))
}
\arguments{
\item{files}{files to be archived}
}
\description{
Intended to be called from \code{\link{to_renv}} as an intermediate step
which saves the existing saproj settings before removing the saproj-specified
.Rprofile and snapshot-library.csv
}
\seealso{
Other functions for migrating to renv: \code{\link{to_renv}}
}
\concept{functions for migrating to renv}
