% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{archive_saproj}
\alias{archive_saproj}
\title{Archive the saproj settings in an "saproj_archive" folder}
\usage{
archive_saproj(
  path = getwd(),
  files_to_move = c(".Rprofile", "snapshot-library.csv")
)
}
\arguments{
\item{path}{file path to project (defaults to working directory)}

\item{files_to_move}{files to be archived}
}
\description{
Archive the saproj settings in an "saproj_archive" folder
}
\seealso{
Other functions for migrating to renv: 
\code{\link{install_saproj_packages}()}
}
\concept{functions for migrating to renv}
