% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-library.R
\name{compare_repo_snapshot}
\alias{compare_repo_snapshot}
\title{Compare snapshot to packages available in repository (helper function)}
\usage{
compare_repo_snapshot(repos = getOption("repos"))
}
\arguments{
\item{repos}{character: The repository searched. The default will likely 
come from MRAN (see 'https://mran.microsoft.com/timemachine' for details).}
}
\description{
This is a helper function for use in \code{\link{restore_library}}. 
It produces a table comparing snapshot package versions with versions
available in the repository (see 'getOption("repos")' for the repository url)
}
\examples{
compare_repo_snapshot
}
\seealso{
Other helper functions for checking project libraries: \code{\link{compare_library_snapshot}}
}
\keyword{internal}
