% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_help.R
\name{dir_create}
\alias{dir_create}
\title{Make a directory (helper function)}
\usage{
dir_create(path, projdir = NULL, showWarnings = FALSE, ...)
}
\arguments{
\item{path}{a character vector containing a single path name}

\item{projdir}{character: Subdirectory to place folders/files, intended for
testing. If NULL, the working directory will be used}

\item{showWarnings}{logical; should the warnings on failure be shown?}

\item{...}{Other parameters passed to \code{\link[base]{dir.create}}}
}
\description{
A helper function for making directories. A variation on
\code{\link[base]{dir.create}} that includes an extra parameter, projdir
}
\examples{
dir_create("test-dir")
dir_create("test-dir", projdir = "proj-dir")
}
\seealso{
Other helper functions for making directories and files: \code{\link{file_copy}},
  \code{\link{make_dat}}
}
