% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_help.R
\name{file_copy}
\alias{file_copy}
\title{Copy a file (helper function)}
\usage{
file_copy(from, to, projdir = NULL, ...)
}
\arguments{
\item{from}{character: file path for origin file}

\item{to}{character: file path for destination file}

\item{projdir}{character: Subdirectory to place folders/files, intended for
testing. If NULL, the working directory will be used}

\item{...}{Other parameters passed to \code{\link[base]{file.copy}}}
}
\description{
This is a helper function for copying files. It is a variation on
\code{\link[base]{file.copy}} that includes an extra parameter, projdir
}
\examples{
file.create("test")
saproj::file_copy("test", "test2")
saproj::file_copy("test", "test2", projdir = "test-dir")
}
\seealso{
Other helper functions for making directories and files: \code{\link{dir_create}},
  \code{\link{make_dat}}
}

