% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_library_path}
\alias{get_library_path}
\title{Get full library path (not exported - internal saproj only)}
\usage{
get_library_path(project_library, inherit_library = FALSE)
}
\arguments{
\item{project_library}{character: Name of project library to use for project packages.
This is required in order to encourage reproducibility and portability.}

\item{inherit_library}{logical: If TRUE, will use an existing project library. 
The default is set to FALSE to avoid accidentally using an existing library.}
}
\value{
Returns the full path to project_library
}
\description{
This is a helper function for use in new_project(), setup_project() and 
update_project(). It gets the full library path and stops function 
execution if the project_library exists (unless 'inherit_library = TRUE'). 
This is intended to avoid a situation where the user accidently inherits 
an existing library.
}
\examples{
get_library_path("example-existing-project")
}
\seealso{
Other internal helper functions: \code{\link{allow_outcome}}
}
\keyword{internal}
