% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{install_packages}
\alias{install_packages}
\title{Install Packages from Repositories or Local Files}
\usage{
install_packages(pkgs, ...)
}
\arguments{
\item{pkgs}{character vector of the names of packages whose
    current versions should be downloaded from the repositories.

    If \code{repos = NULL}, a character vector of
    file paths of \file{.zip} files containing binary builds of
    packages.  (\code{http://} and \code{file://} URLs are also accepted
    and the files will be downloaded and installed from local copies.)
    Source directories or file paths or URLs of archives may be
    specified with \code{type = "source"}, but some packages need
    suitable tools installed (see the \sQuote{Details} section).
#ifdef unix not active
    If this is missing or a zero-length character vector, a listbox of
    available packages is presented where possible in an interactive \R
    session.
  }

\item{...}{Arguments passed to \code{\link[utils]{install.packages}}}
}
\description{
This is a wrapper for \code{\link[utils]{install.packages}} that creates a record
of package installations. It always installs to the first element of 
\code{\link[base]{.libPaths}}, which should correspond to either a project-specific 
or user library. You can check this by running .libPaths() in the console.
}
\examples{
Install a CRAN package
saproj::install_packages("acs")

Install a local package
saproj::install_packages("pkg/salic_0.3.4.zip", repos = NULL, type = "win.binary")
}
\seealso{
Other functions for working with R packages: \code{\link{view_packages}}
}
