% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{install_saproj_packages}
\alias{install_saproj_packages}
\title{Install previous saproj packages into renv library}
\usage{
install_saproj_packages(
  snapshot = file.path("saproj_archive", "snapshot-library.csv"),
  exclude = c("sadash@1.0.4", "urbnmapr@0.0.0.9002", "sf@0.6-3")
)
}
\arguments{
\item{snapshot}{path to snapshot-library.csv used by saproj}

\item{exclude}{optionally exclude specified packages. Useful if
\code{\link[renv]{install}} can't find the specified package and produces a cryptic error.}
}
\description{
To be called following \code{\link{archive_saproj}} and renv \code{\link[renv]{init}}.
If you get a cryptic error such as "Cannot open URL...", you can excluded the 
offending package-version with the exclude argument
}
\seealso{
Other functions for migrating to renv: 
\code{\link{archive_saproj}()}
}
\concept{functions for migrating to renv}
