% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_help.R
\name{make_dat}
\alias{make_dat}
\title{Make a data directory (helper function)}
\usage{
make_dat(dat_path)
}
\arguments{
\item{dat_path}{character: file path for the directory where the data
directory will be written to}
}
\description{
A helper function for use in saproj new. Assumes the base project directory is
one level above the R project directory
}
\examples{
saproj::make_dat("D:/test")
}
\seealso{
Other helper.functions.for.making.directories.and.files: \code{\link{dir_create}},
  \code{\link{file_copy}}
}

