% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{new}
\alias{new}
\title{Populate a project with default directories and files}
\usage{
new(dat = FALSE, dat2 = FALSE, dat_path = "D:/SA/Data",
  dat_path2 = "D:/SA/Data2", projdir = NULL, generic = TRUE)
}
\arguments{
\item{dat}{logical: If TRUE, make directory in dat_path}

\item{dat2}{logical: If TRUE, make directory in dat_path2}

\item{dat_path}{character: file path to optional dat directory}

\item{dat_path2}{character: file path to optional dat2 directory}

\item{projdir}{character: Subdirectory to place folders/files, intended for
testing. If NULL, the working directory will be used}

\item{generic}{logical: If TRUE, generic README & run.R files are also
created. Otherwise, these won't be created, allowing for automation of
custom file creation}
}
\description{
Populate a project with default directories and files
}
\examples{
saproj::new(projdir = "test-dir")
}
\seealso{
Other functions for making directories and files: \code{\link{ref}},
  \code{\link{section}}
}

