% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{new_project}
\alias{new_project}
\title{Populate a project with default directories and files}
\usage{
new_project(proj_library, inherit_library = FALSE, projdir = NULL)
}
\arguments{
\item{proj_library}{character: Name of project library}

\item{inherit_library}{logical: If TRUE, will use an existing project library}

\item{projdir}{character: Subdirectory to place folders/files, intended for
testing. If NULL, the working directory will be used.}
}
\description{
Populate a project with default directories and files
}
\examples{
saproj::new_project("some-project-name", projdir = "test-dir")
}
\seealso{
Other functions for setting up projects: \code{\link{new_section}}
}
