% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{ref}
\alias{ref}
\title{Copy selected package (zipfile) to a directory}
\usage{
ref(pkg, pkg_dir = "D:/SA/Project_DK/R", version = NULL, ref_dir = "ref",
  projdir = NULL)
}
\arguments{
\item{pkg}{character: Name of package to copy}

\item{pkg_dir}{character: path to directory holding package zipfiles}

\item{version}{character: Optional version of package to copy. If NULL,
the most recent version will be used}

\item{ref_dir}{character: Name of directory where packages will be stored}

\item{projdir}{character: Subdirectory to place folders/files, intended for
testing. If NULL, the working directory will be used}
}
\description{
Intended to store install files for packages that aren't on CRAN.
This function automates storing these files for convenience
}
\examples{
saproj::new(projdir = "test-dir")
saproj::ref("lichist", projdir = "test-dir")
saproj::ref("lichist", version = "0.1.5", projdir = "test-dir")
}
\seealso{
Other functions for making directories and files: \code{\link{new}},
  \code{\link{section}}
}

