% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-library.R
\name{restore_library}
\alias{restore_library}
\title{Restore a project library using a snapshot}
\usage{
restore_library(proj_libpath = .libPaths()[1], override_version = FALSE)
}
\arguments{
\item{proj_libpath}{character: The location of the project library 
(note: changing this argument is not recommended)}

\item{override_version}{logical: If TRUE, uses the version available in getOption("repos")
regardless of the version specified in the snapshot.}
}
\description{
This is intended to be run when a project needs to be rerun or edited
on a different computer. It installs the packages listed in 'project-snapshot.csv'.
}
\examples{
restore_library()
}
\seealso{
Other functions for maintaining project package libraries: \code{\link{snapshot_library}},
  \code{\link{view_library}}
}
