% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{rmd}
\alias{rmd}
\title{Source an R script using an Rmd file}
\usage{
rmd(rmd_file, indir, src_file, title = "Sourced Code")
}
\arguments{
\item{rmd_file}{character: path to Rmd file used for sourcing}

\item{indir}{character: path to code sub-directory that contains the R script}

\item{src_file}{character: name of the R script to source}

\item{title}{character: Optional title to put at the top of the output file}
}
\description{
Note that this is a deprecated feature. It's better to simply write your analysis
in R Markdown
}
\examples{
# (later) maybe include sample R scripts for sourcing
rmd()
}

