% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{section}
\alias{section}
\title{Create a new sub-directory for R scripts, docs, etc.}
\usage{
section(title, projdir = NULL)
}
\arguments{
\item{title}{character: Name of sub-directory to create}

\item{projdir}{character: Subdirectory to place folders/files, intended for
testing. If NULL, the working directory will be used}
}
\description{
Create a new sub-directory for R scripts, docs, etc.
}
\examples{
saproj::new(projdir = "test-dir")
saproj::section("1-first-section", projdir = "test-dir")
}
\seealso{
Other functions for making directories and files: \code{\link{new}},
  \code{\link{ref}}
}

