% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{to_renv}
\alias{to_renv}
\title{Convert saproj project settings for use with package renv}
\usage{
to_renv(update_r = FALSE, update_packages = FALSE,
  files = c(".Rprofile", "snapshot-library.csv"))
}
\arguments{
\item{update_r}{If TRUE, your default R version will be used for renv. If you 
update to a new version, there is a small possibility that breaking changes 
will be introduced.}

\item{update_packages}{If TRUE, will use new package versions instead of those
specified in snapshot-library.csv. The chances of breaking changes are greater
than those related to R version upates.}

\item{files}{files to be archived}
}
\description{
This is intended for migration to a better dependency management system (renv).
The existing saproj settings will be archived in an "saproj" folder using 
\code{\link{archive_project}} in case they need to be retrieved in the future. 
An renv lockfile will be created for use with renv::restore().
}
\seealso{
Other functions for migrating to renv: \code{\link{archive_project}}
}
\concept{functions for migrating to renv}
