% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{update_project}
\alias{update_project}
\title{Update project R version or package library}
\usage{
update_project(project_library = NULL, r_version = NULL,
  inherit_library = FALSE)
}
\arguments{
\item{project_library}{character: New project library to use. Defaults
to no change.}

\item{r_version}{character: New R version to use (e.g., "3.4.3"). Defaults
to no change.}

\item{inherit_library}{logical: If TRUE, will use an existing project library. 
The default is set to FALSE to avoid accidentally using an existing library.}
}
\description{
This allows the 2 basic reproducibility parameters to be updated. This
would be useful (for example) if you want to adapt code/data of an 
existing project for a new project.
}
\examples{
# saproj::update_project(project_library = "new-project-name")
# saproj::update_project(r_version = "3.4.3")
}
\seealso{
Other functions for setting up projects: \code{\link{new_project}},
  \code{\link{new_section}}, \code{\link{setup_project}},
  \code{\link{view_projects}}
}
\concept{functions for setting up projects}
