% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-library.R
\name{view_library}
\alias{view_library}
\title{Convenience function to look at a package library}
\usage{
view_library(library_path = .libPaths()[1])
}
\arguments{
\item{library_path}{character: path to package library}
}
\description{
This is a quick way of looking at packages installed, particularly useful for a 
project-specific library.
}
\examples{
view_library()
}
\seealso{
Other functions for maintaining project package libraries: \code{\link{compare_library_snapshot}},
  \code{\link{snapshot_library}}
}
