% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{view_packages}
\alias{view_packages}
\title{View packages installed in library}
\usage{
view_packages(library_path = .libPaths()[1])
}
\arguments{
\item{library_path}{character: path to package library}
}
\description{
This is a quick way of looking at packages installed, particularly useful for a 
project-specific library.
}
\examples{
view_packages()
}
\seealso{
Other functions for working with R packages: \code{\link{install_packages}}
}
