% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{view_projects}
\alias{view_projects}
\title{List project libraries stored on this computer}
\usage{
view_projects(list_packages = FALSE, all_versions = FALSE)
}
\arguments{
\item{list_packages}{logical: If TRUE, includes a list of packages installed
in each project library.}

\item{all_versions}{logical: If TRUE, include all versions of R installed on 
this machine. Defaults to using only the currently loaded version.}
}
\description{
This is intended to help out if you are wondering what to name a new project
(i.e., the 'project_library' parameter in \code{\link{new_project}}, etc.).
}
\examples{
view_projects()
view_projects(list_packages = TRUE)
view_projects(list_packages = TRUE, all_versions = TRUE)
}
\seealso{
Other functions for setting up projects: \code{\link{new_project}},
  \code{\link{new_section}}, \code{\link{setup_project}},
  \code{\link{update_project}}
}
\concept{functions for setting up projects}
