% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-library.R
\name{compare_library_snapshot}
\alias{compare_library_snapshot}
\title{Convenience function to compare installed project packages to snapshot-library.csv}
\usage{
compare_library_snapshot(proj_libpath = .libPaths()[1])
}
\arguments{
\item{proj_libpath}{character: The location of the project library 
(note: changing this argument is not recommended)}
}
\description{
The results of the comparison inform the recommendation for actions to 
sync the installed library and it's snapshot.
}
\examples{
compare_library_snapshot()
}
\seealso{
Other functions for maintaining project package libraries: \code{\link{restore_library}},
  \code{\link{snapshot_library}},
  \code{\link{view_library}}
}
