% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{new_project}
\alias{new_project}
\title{Populate a project with default directories and files}
\usage{
new_project(project_library, inherit_library = FALSE,
  r_version = paste(R.version$major, R.version$minor, sep = "."))
}
\arguments{
\item{project_library}{character: Name of project library to use for project-specific packages.
This is required in order to encourage reproducibility and portability.}

\item{inherit_library}{logical: If TRUE, will use an existing project library. 
The default is set to FALSE to avoid accidentally using an existing library.}

\item{r_version}{character: R version (e.g., "3.4.3") to use for the new project.
Defaults to currently-used version.}
}
\description{
This also creates a space for project-specific packages (using .Rprofile)
}
\examples{
saproj::new_project("some-project-name")
}
\seealso{
Other functions for setting up projects: \code{\link{new_section}}
}
