% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{new_project}
\alias{new_project}
\title{Populate a project with default directories and files}
\usage{
new_project(dat = FALSE, dat2 = FALSE, dat_path = "D:/SA/Data",
  dat_path2 = "D:/SA/Data2", projdir = NULL)
}
\arguments{
\item{dat}{logical: If TRUE, make directory in dat_path}

\item{dat2}{logical: If TRUE, make directory in dat_path2}

\item{dat_path}{character: file path to optional dat directory}

\item{dat_path2}{character: file path to optional dat2 directory}

\item{projdir}{character: Subdirectory to place folders/files, intended for
testing. If NULL, the working directory will be used}
}
\description{
Populate a project with default directories and files
}
\examples{
saproj::new_project(projdir = "test-dir")
}
\seealso{
Other functions for making directories and files: \code{\link{new_section}},
  \code{\link{ref}}
}

