% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{new_section}
\alias{new_section}
\title{Create a new sub-directory for R code, data, and output}
\usage{
new_section(section_title, recursive = TRUE)
}
\arguments{
\item{section_title}{character: Name of sub-directory to create}

\item{recursive}{logical. Should elements of the path other than the
    last be created?  If true, like the Unix command \command{mkdir -p}.}
}
\description{
This is a convenience function to start a new code 
section by including corresponding section folders in "data/" and "out/".
}
\examples{
saproj::new_project()
saproj::new_section("1-first-section")
}
\seealso{
Other functions for setting up projects: \code{\link{new_project}}
}
