% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{new_section}
\alias{new_section}
\title{Create a new sub-directory for R scripts, docs, etc.}
\usage{
new_section(title, projdir = NULL)
}
\arguments{
\item{title}{character: Name of sub-directory to create}

\item{projdir}{character: Subdirectory to place folders/files, intended for
testing. If NULL, the working directory will be used}
}
\description{
Create a new sub-directory for R scripts, docs, etc.
}
\examples{
saproj::new_project(projdir = "test-dir")
saproj::new_section("1-first-section", projdir = "test-dir")
}
\seealso{
Other functions.for.making.directories.and.files: \code{\link{new_project}},
  \code{\link{ref}}
}

