% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-library.R
\name{restore_library}
\alias{restore_library}
\title{Restore a snapshot to a package library}
\usage{
restore_library(proj_libpath = .libPaths()[1], override_version = FALSE)
}
\arguments{
\item{proj_libpath}{character: The location of the project library 
(note: changing this argument is not recommended)}
}
\description{
This installs only those snapshot packages not included in the library.
}
\examples{
restore_library()
}
\seealso{
Other functions for maintaining project package libraries: \code{\link{compare_library_snapshot}},
  \code{\link{snapshot_library}},
  \code{\link{view_library}}
}
