% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-library.R
\name{restore_library}
\alias{restore_library}
\title{Restore a project library using a snapshot}
\usage{
restore_library(proj_libpath = .libPaths()[1],
  repos = getOption("repos"), override_version = FALSE,
  use_devtools = FALSE, devtools_repo = "https://cran.rstudio.com")
}
\arguments{
\item{proj_libpath}{character: The location of the project library 
(note: changing this value is not recommended)}

\item{repos}{character: The repository searched. The default will likely 
come from MRAN (see 'https://mran.microsoft.com/timemachine' for details).}

\item{override_version}{logical: If TRUE, uses the version available in repos
regardless of the version specified in the snapshot (and updates the snapshot 
accordingly).}

\item{use_devtools}{logical: If TRUE, uses \code{\link[devtools]{install_version}}.
This may require additional configuration of your computer (installing RTools).
RStudio might open a prompt to begin RTools installation if you select 'OK'.}

\item{devtools_repo}{character: Repository to use when use_devtools = TRUE}
}
\description{
This is intended to be run when a project analysis needs to be re-run or edited
on a different computer. It installs the packages listed in 'project-snapshot.csv'.
}
\examples{
# saproj::restore_library()
}
\seealso{
Other functions for maintaining project package libraries: \code{\link{snapshot_library}},
  \code{\link{view_library}}
}
\concept{functions for maintaining project package libraries}
