% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-library.R
\name{snapshot_library}
\alias{snapshot_library}
\title{Create a snapshot (csv) of packages installed in project library}
\usage{
snapshot_library(proj_libpath = .libPaths()[1], replace_snapshot = FALSE)
}
\arguments{
\item{proj_libpath}{character: The location of the project library 
(note: changing this value is not recommended)}

\item{replace_snapshot}{logical: If TRUE, any existing snapshot will be removed
and a snapshot will be created based on the current project library
(if any packages are installed).}
}
\description{
The snapshot improves project portability by storing details about installed
packages. The project library can be restored on a different machine using 
\code{\link{restore_library}}
}
\examples{
saproj::snapshot_library()
}
\seealso{
Other functions for maintaining project package libraries: \code{\link{restore_library}},
  \code{\link{view_library}}
}
