% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-library.R
\name{snapshot_library}
\alias{snapshot_library}
\title{Create a snapshot (csv) of packages installed in project library}
\usage{
snapshot_library(proj_libpath = .libPaths()[1], resolve_conflicts = FALSE)
}
\description{
The snapshot improves project portability by storing details about installed
packages. The project library can then be restored using \code{\link{restore_library}}
}
