% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-library.R
\name{view_library}
\alias{view_library}
\title{List packages installed in library}
\usage{
view_library(library_path = .libPaths()[1])
}
\arguments{
\item{library_path}{character: Path to package library. Defaults to the top-level
library (typically a project-specific library if \code{\link{new_project}} was used
to initialize a the project).}
}
\description{
This is a quick way of looking at packages installed, particularly useful for a 
project-specific library.
}
\examples{
# saproj::view_library()
}
\seealso{
Other functions for maintaining project package libraries: \code{\link{restore_library}},
  \code{\link{snapshot_library}}
}
\concept{functions for maintaining project package libraries}
